moved interfaces in separate files, moved mock data to corresponding directory and UPPERCASED static data variables
