+++ src/App.tsx
@@ -1,13 +1,13 @@
 import { FC } from 'react';
 import { Footer } from '@/components/layout/Footer';
 import { Header } from '@/components/layout/Header';
-import { Homepage } from '@/pages/Homepage';
+import { HomePage } from '@/pages/HomePage';
 
 export const App: FC = () => (
   <div className="flex h-dvh min-h-[250px] w-screen min-w-[300px] max-w-[1600px] flex-col overflow-hidden font-main text-text">
     <Header />
     <div className="no-scrollbar grow overflow-y-scroll">
-      <Homepage />
+      <HomePage />
       <Footer />
     </div>
   </div>
