+++ src/components/homepage/CategoriesSection.tsx
@@ -1,50 +1,8 @@
 import { FC } from 'react';
-import { CategoryCard, Category } from '@/components/common/CategoryCard';
+import { CategoryCard } from '@/components/common/CategoryCard';
+import { CATEGORIES } from '@/data/categories';
 import { BREAKPOINTS } from '@/lib/use-breakpoint';
 import { useBreakpoint } from 'use-breakpoint';
-
-const categories: Category[] = [
-  {
-    title: 'Art',
-    icon: '/categories/icons/art.svg',
-    preview: '/categories/previews/art.jpg',
-  },
-  {
-    title: 'Collectibles',
-    icon: '/categories/icons/collectibles.svg',
-    preview: '/categories/previews/collectibles.jpg',
-  },
-  {
-    title: 'Music',
-    icon: '/categories/icons/music.svg',
-    preview: '/categories/previews/music.jpg',
-  },
-  {
-    title: 'Photography',
-    icon: '/categories/icons/photography.svg',
-    preview: '/categories/previews/photography.jpg',
-  },
-  {
-    title: 'Video',
-    icon: '/categories/icons/video.svg',
-    preview: '/categories/previews/video.jpg',
-  },
-  {
-    title: 'Utility',
-    icon: '/categories/icons/utility.svg',
-    preview: '/categories/previews/utility.jpg',
-  },
-  {
-    title: 'Sport',
-    icon: '/categories/icons/sport.svg',
-    preview: '/categories/previews/sport.jpg',
-  },
-  {
-    title: 'Virtual Worlds',
-    icon: '/categories/icons/virtual-worlds.svg',
-    preview: '/categories/previews/virtual-worlds.jpg',
-  },
-];
 
 export const CategoriesSection: FC = () => {
   const { minWidth } = useBreakpoint(BREAKPOINTS, 'xs');
@@ -54,7 +12,7 @@ export const CategoriesSection: FC = () => {
       <h1 className="mb-9 text-3xl font-semibold lg:mb-[3.75rem] lg:text-4xl">Browse Categories</h1>
 
       <div className="grid grid-cols-2 gap-5 sm:grid-cols-3 sm:gap-7 md:grid-cols-4 lg:grid-cols-3 xl:grid-cols-4">
-        {categories.map((category) => (
+        {CATEGORIES.map((category) => (
           <CategoryCard
             key={category.title}
             category={category}
