+++ src/data/categories.ts
@@ -1,0 +1,44 @@
+import { Category } from '@/types/Category';
+
+export const CATEGORIES: Category[] = [
+  {
+    title: 'Art',
+    icon: '/categories/icons/art.svg',
+    preview: '/categories/previews/art.jpg',
+  },
+  {
+    title: 'Collectibles',
+    icon: '/categories/icons/collectibles.svg',
+    preview: '/categories/previews/collectibles.jpg',
+  },
+  {
+    title: 'Music',
+    icon: '/categories/icons/music.svg',
+    preview: '/categories/previews/music.jpg',
+  },
+  {
+    title: 'Photography',
+    icon: '/categories/icons/photography.svg',
+    preview: '/categories/previews/photography.jpg',
+  },
+  {
+    title: 'Video',
+    icon: '/categories/icons/video.svg',
+    preview: '/categories/previews/video.jpg',
+  },
+  {
+    title: 'Utility',
+    icon: '/categories/icons/utility.svg',
+    preview: '/categories/previews/utility.jpg',
+  },
+  {
+    title: 'Sport',
+    icon: '/categories/icons/sport.svg',
+    preview: '/categories/previews/sport.jpg',
+  },
+  {
+    title: 'Virtual Worlds',
+    icon: '/categories/icons/virtual-worlds.svg',
+    preview: '/categories/previews/virtual-worlds.jpg',
+  },
+];
