+++ src/data/collections.ts
@@ -1,0 +1,46 @@
+import { Collection } from '@/types/Collection';
+
+export const COLLECTIONS: Collection[] = [
+  {
+    title: 'DSGN Animals',
+    creator: {
+      image: '/creators/mrfox.jpg',
+      name: 'MrFox',
+      totalSales: 34.53,
+    },
+    previews: [
+      '/nfts/dsgn-animals-1-by-mrfox.jpg',
+      '/nfts/dsgn-animals-2-by-mrfox.jpg',
+      '/nfts/dsgn-animals-3-by-mrfox.jpg',
+    ],
+    size: 1028,
+  },
+  {
+    title: 'Magic Mushrooms',
+    creator: {
+      image: '/creators/shroomie.jpg',
+      name: 'Shroomie',
+      totalSales: 34.53,
+    },
+    previews: [
+      '/nfts/magic-mushrooms-1-by-shroomie.jpg',
+      '/nfts/magic-mushrooms-2-by-shroomie.jpg',
+      '/nfts/magic-mushrooms-3-by-shroomie.jpg',
+    ],
+    size: 1028,
+  },
+  {
+    title: 'Disco Machines',
+    creator: {
+      image: '/creators/bekind2robots.jpg',
+      name: 'BeKind2Robots',
+      totalSales: 34.53,
+    },
+    previews: [
+      '/nfts/disco-machines-1-by-bekind2robots.jpg',
+      '/nfts/disco-machines-2-by-bekind2robots.jpg',
+      '/nfts/disco-machines-3-by-bekind2robots.jpg',
+    ],
+    size: 1028,
+  },
+];
