+++ src/data/creators.ts
@@ -1,0 +1,64 @@
+import { Creator } from '@/types/Creator';
+
+export const CREATORS: Creator[] = [
+  {
+    name: 'Keepitreal',
+    image: '/creators/keepitreal.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'DigiLab',
+    image: '/creators/digilab.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'GravityOne',
+    image: '/creators/gravityone.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'Juanie',
+    image: '/creators/juanie.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'BlueWhale',
+    image: '/creators/bluewhale.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'MrFox',
+    image: '/creators/mrfox.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'Shroomie',
+    image: '/creators/shroomie.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'Robotica',
+    image: '/creators/robotica.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'RustyRobot',
+    image: '/creators/rustyrobot.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'Animakid',
+    image: '/creators/animakid.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'Dotgu',
+    image: '/creators/dotgu.jpg',
+    totalSales: 34.53,
+  },
+  {
+    name: 'Ghiblier',
+    image: '/creators/ghiblier.jpg',
+    totalSales: 34.53,
+  },
+];
