+++ src/pages/Homepage.tsx
@@ -8,7 +8,7 @@ import { AuctionSection } from '@/components/homepage/AuctionSection';
 import { HowItWorksSection } from '@/components/homepage/HowItWorksSection';
 import { WeeklyDigestSection } from '@/components/homepage/WeeklyDigestSection';
 
-export const Homepage: FC = () => (
+export const HomePage: FC = () => (
   <main role="main">
     <WelcomeSection />
     <CollectionsSection />
