+++ src/types/Category.ts
@@ -1,0 +1,5 @@
+export interface Category {
+  icon: string;
+  title: string;
+  preview: string;
+}
