+++ src/types/Collection.ts
@@ -1,0 +1,8 @@
+import { Creator } from '@/types/Creator';
+
+export interface Collection {
+  title: string;
+  creator: Creator;
+  previews: string[];
+  size: number;
+}
