+++ src/types/Creator.ts
@@ -1,0 +1,5 @@
+export interface Creator {
+  image: string;
+  name: string;
+  totalSales: number;
+}
