+++ src/types/Nft.ts
@@ -1,0 +1,9 @@
+import { Creator } from '@/types/Creator';
+
+export interface Nft {
+  image: string;
+  title: string;
+  creator: Creator;
+  price: number;
+  highestBid: number;
+}
